+++ src/app/home/components/ExperimentPostContainer/ExperimentPostContainer.tsx
@@ -1,7 +1,5 @@
 'use client';
 
-import { useEffect, useState } from 'react';
-
 import ExperimentPostCardListContainer from './ExperimentPostCardListContainer/ExperimentPostCardListContainer';
 import {
   filterWrapper,
@@ -10,57 +8,22 @@ import {
   recruitCheckLabel,
 } from './ExperimentPostContainer.css';
 import FilterContainer from './FilterContainer/FilterContainer';
-import { calculateAgeFromBirthDate, filterParticipantInfo } from '../../home.utils';
+import useExperimentFilters from '../../hooks/useExperimentFilters';
 import useUserInfo from '../../hooks/useUserInfo';
 
-import { ExperimentPostListFilters } from '@/apis/post';
 import JoinCheckbox from '@/app/join/components/JoinCheckboxContainer/JoinCheckbox/JoinCheckbox';
 import Icon from '@/components/Icon';
 
 const ExperimentPostContainer = () => {
   const { userInfo, isLoading: isUserInfoLoading } = useUserInfo();
-  const participantInfo = filterParticipantInfo(userInfo);
 
-  const [filters, setFilters] = useState<ExperimentPostListFilters>(
-    {} as ExperimentPostListFilters,
-  );
-
-  const isRecruiting = filters.recruitStatus === 'OPEN';
-
-  const handleFilterChange = (key: string, value: string | string[] | number | boolean | null) => {
-    setFilters((prev) => ({
-      ...prev,
-      [key]: value,
-    }));
-  };
-
-  const handleChange = () => {
-    const toggleChecked = isRecruiting ? 'ALL' : 'OPEN';
-    handleFilterChange('recruitStatus', toggleChecked);
-  };
-
-  const handleResetFilter = () => {
-    setFilters({
-      recruitStatus: 'ALL',
-    });
-  };
-
-  // 참여자 성별/나이 자동 필터링
-  useEffect(() => {
-    if (participantInfo) {
-      setFilters((prev) => ({
-        ...prev,
-        recruitStatus: 'ALL',
-        gender: participantInfo.gender,
-        age: calculateAgeFromBirthDate(participantInfo.birthDate),
-      }));
-    } else {
-      setFilters((prev) => ({
-        ...prev,
-        recruitStatus: 'ALL',
-      }));
-    }
-  }, [participantInfo]);
+  const {
+    filters,
+    isRecruiting,
+    handleFilterChange,
+    handleToggleRecruitStatus,
+    handleResetFilter,
+  } = useExperimentFilters(userInfo);
 
   return (
     <div className={postContainerLayout}>
@@ -78,7 +41,7 @@ const ExperimentPostContainer = () => {
           labelClassName={recruitCheckLabel}
           label="모집 중인 공고만 보기"
           isChecked={isRecruiting}
-          onChange={handleChange}
+          onChange={handleToggleRecruitStatus}
           isArrow={false}
           emptyCheckIcon={<Icon icon="CheckSquareFill" />}
         />
